parsing single value mandatory elements
