+++ tests/test.metadata/abstract.xml
@@ -1,0 +1,16 @@
+<gmd:MD_Metadata xmlns:gmd="http://www.isotc211.org/2005/gmd" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:gco="http://www.isotc211.org/2005/gco" xmlns:geonet="http://www.fao.org/geonetwork" xmlns:xlink="http://www.w3.org/1999/xlink" xmlns:gmx="http://www.isotc211.org/2005/gmx" xmlns:eamp="http://environment.data.gov.uk/eamp" xmlns:csw="http://www.opengis.net/cat/csw/2.0.2" xsi:schemaLocation="http://www.isotc211.org/2005/gmx http://eden.ign.fr/xsd/isotc211/isofull/20090316/gmx/gmx.xsd">
+  <gmd:identificationInfo>
+    <gmd:MD_DataIdentification id="BGS-13603180">
+      <gmd:abstract>
+        <gco:CharacterString>The data shows polygonal and selected linear geological information, sourced from published 
+          BGS 1:625 000 scale maps of Great Britain. 
+          However, geological units are identified using the most up-to-date nomenclature that may differ from that on the 
+          printed maps. The maps are generally based on published material at 1:50 000 scale and compiled using techniques of 
+          selection, generalisation and exaggeration. The geology is fitted to a relevant topographic base at the time of 
+          production. Full UK coverage is available. The data is available in vector format. BGS licensing terms and 
+          conditions apply to external use of the data.
+        </gco:CharacterString>
+      </gmd:abstract>
+    </gmd:MD_DataIdentification>
+  </gmd:identificationInfo>
+</gmd:MD_Metadata>
